refact: 도메인 기준으로 파일 분리
