+++ main.py
@@ -1,11 +1,14 @@
 from fastapi import FastAPI
-from src.router import logRouter, eventRouter, widgetRouter, planRouter
+from src.router import workerRouter, cellRouter, robotArmRouter, amrRouter, eventRouter, widgetRouter, planRouter
 from fastapi.middleware.cors import CORSMiddleware
 from src.socket import websocketHandler
 import uvicorn  
 
 app = FastAPI()
-app.include_router(logRouter.api, prefix='/log')
+app.include_router(workerRouter.api, prefix='/worker')
+app.include_router(cellRouter.api, prefix='/cell')
+app.include_router(robotArmRouter.api, prefix='/robot-arm')
+app.include_router(amrRouter.api, prefix='/amr')
 app.include_router(eventRouter.api, prefix='/evnet')
 app.include_router(widgetRouter.api, prefix='/widget')
 app.include_router(planRouter.api, prefix='/plan')
